:recycle: Search structure refactor
