+++ aksel.nav.no/website/components/layout/header/search/providers/index.ts
@@ -1,0 +1,2 @@
+export * from "./SearchProvider";
+export * from "./SearchResultProvider";
